+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod libs;
 
+pub use crate::libs::loc::*;
 pub use crate::libs::nt::*;