+++ src/libs/mod.rs
@@ -1,1 +1,2 @@
+pub mod loc;
 pub mod nt;