+++ tests/cli_index.rs
@@ -28,3 +28,28 @@ fn command_gz() -> anyhow::Result<()> {
     tempdir.close()?;
     Ok(())
 }
+
+#[test]
+fn command_range() -> anyhow::Result<()> {
+    let mut cmd = Command::cargo_bin("hnsm")?;
+    let output = cmd
+        .arg("range")
+        .arg("tests/index/final.contigs.fa.gz")
+        .arg("k81_130")
+        .arg("k81_130:11-20")
+        .arg("k81_170:304-323")
+        .arg("k81_170(-):1-20")
+        .arg("k81_158:70001-70020")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 10);
+    assert!(stdout.contains(">k81_130\nAGTTTCAACT"));
+    assert!(stdout.contains(">k81_130:11-20\nGGTGAATCAA\n"));
+    assert!(stdout.contains(">k81_170:304-323\nAGTTAAAAACCTGATTTATT\n"));
+    assert!(stdout.contains(">k81_170(-):1-20\nAATAAATCAGGTTTTTAACT\n"));
+    assert!(stdout.contains(">k81_158:70001-70020\nTGGCTATAACCTAATTTTGT\n"));
+
+    Ok(())
+}