+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Index Subcommands
     * gz
+    * range
 
 ## 0.1.0 - 2024-07-11